pkg: manifests: cleanup useless fields
Let's remove fields which doesn't make sense
as render target, because they are managed by the cluster
and thus are ignored anyway when we submit the manifests,
like status and creationTimestamps.
